requirements.txt
streamlit-chat
langchain
qdrant_client